+++ src/main/java/nextstep/courses/domain/Course.java
@@ -2,6 +2,7 @@ package nextstep.courses.domain;
 
 import java.io.File;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 
 public class Course {
@@ -15,19 +16,7 @@ public class Course {
 
     private LocalDateTime updatedAt;
 
-    private LocalDateTime startedAt;
-
-    private LocalDateTime endedAt;
-
-    private File coverImage;
-
-    private boolean isFree;
-
-    private Status status;
-
-    private int currentStudents;
-
-    private int maxStudents;
+    private List<Session> sessions;
 
     public Course() {
     }
@@ -42,12 +31,7 @@ public class Course {
         this.creatorId = creatorId;
         this.createdAt = createdAt;
         this.updatedAt = updatedAt;
-        this.startedAt = null;
-        this.endedAt = null;
-        this.isFree = true;
-        this.status = Status.preparing;
-        this.currentStudents = 0;
-        this.maxStudents = 0;
+        sessions = new ArrayList<>();
     }
 
     public String getTitle() {
@@ -62,55 +46,13 @@ public class Course {
         return createdAt;
     }
 
-    public void patchTerms(LocalDateTime startedAt, LocalDateTime endedAt){
-        this.startedAt = startedAt;
-        this.endedAt = endedAt;
-    }
-    public List<LocalDateTime> terms() {
-        return List.of(startedAt, endedAt);
+    public void addSession(LocalDateTime startedAt, LocalDateTime endedAt, boolean isFree, Status status, int currentStudents, int maxStudents) {
+        Session newSession = new Session(startedAt, endedAt, isFree, status, currentStudents, maxStudents);
+        this.sessions.add(newSession);
     }
 
-    public void patchIsFree(boolean input) {
-        this.isFree = input;
-    }
-
-    public boolean isFree() {
-        return isFree;
-    }
-
-    public void registerCoverImage(File file) {
-        this.coverImage = file;
-    }
-
-    public File coverImage() {
-        return this.coverImage;
-    }
-
-    public void patchStatus(String status) {
-        this.status = Status.valueOf(status);
-    }
-    public boolean isOpening() {
-        return status == Status.opening;
-    }
-
-    public void registerMaxStudents(int count) {
-        this.maxStudents = count;
-    }
-
-    public void enrolement() {
-        if (isOpening()) {
-            throw new RuntimeException("해당 강의는 모집중이 아닙니다.");
-        }
-
-        if (currentStudents >= maxStudents) {
-            throw new RuntimeException("정원이 가득찼습니다.");
-        }
-
-        currentStudents++;
-    }
-
-    public int currentStudents() {
-        return currentStudents;
+    public List<Session> sessionList(){
+        return sessions;
     }
 
     @Override
